Sync jparse/ from jparse repo
Synced jparse/ from the jparse repo (https://github.com/xexyl/jparse/).

This adds a new utility function (that is tested in make test via the
util_test.c which comes from util.c) that opens and parses a path as a
JSON file (by use of the open_dir_file() function), returning NULL if
invalid JSON. This will be used in the required updates to chkentry(1)
post IOCCC28.